Backend.sanitizing.valmpani and Chats (#29)
* fix: infra: added redis variables to sample

* feat: backend: pong, stats, tournaments are sanitized

* feat: api: websocket client for frontend

* feat: frontend: Usable chat interface - not finished

* fix: backend.chat: uniform serialization of message

* feat: frontend: Better desing for chats and active users -> broken primary color

* feat: backend: messages are marked as read after frontend event

* fix: frontend: double modal in the end of local game

* feat: frontend: unread messages working for multiple chats

* fix: backend: total unread messages

* fix: frontend: unread message badge in barlayout

* fix: chats: sync operation from async context and check for sender in barlayout for badge

* adds consumer handler for active friend ids request message

* feat: frontend: user status in users page

* resolved conflict of channel_layer group in tournament consumer

* fix: tournament: fix for avatar functional component, user status and unread messages

---------

+++ srcs/backend/transcendence_backend/pong/views.py
@@ -19,13 +19,15 @@ class PongView(View):
         @param user: User
         @return: JsonResponse with game schema
         """
-        type = str(request.GET.get('type', ''))
-        skip = int(request.GET.get('skip', 0))
-        limit = int(request.GET.get('limit', 10))
-        me = request.GET.get('me', False)
-        if type == '':
-            return JsonResponse({'Error': 'Invalid type'}, status=400)
         try:
+            type = str(request.GET.get('type', ''))
+            skip = int(request.GET.get('skip', 0))
+            limit = int(request.GET.get('limit', 10))
+            me = request.GET.get('me', False)
+            if skip < 0 or limit < 0:
+                return JsonResponse({'Error': 'Invalid skip or limit'}, status=400)
+            if type == '':
+                return JsonResponse({'Error': 'Invalid type'}, status=400)
             games = PongService.get_games(user, type, skip, limit, me)
             return JsonResponse(games, status=200, safe=False)
         except Exception as e:
@@ -41,6 +43,8 @@ class PongView(View):
         """
         try:
             max_points = int(request.GET.get('max_points', 10))
+            if max_points < 1:
+                return JsonResponse({'Error': 'Invalid max_points'}, status=400)
             game = PongService.create_game(user, max_points)
             return JsonResponse(game, status=201)
         except Exception as e:
@@ -53,18 +57,20 @@ class PongView(View):
         @param user: User
         @return: JsonResponse with game schema
         """
-        if request.body == b'':
-            return JsonResponse({'Error': 'Invalid request'}, status=400)
-        data = json.loads(request.body)
         try:
-            game_id = data.get('game_id', None)
+            if request.body == b'':
+                return JsonResponse({'Error': 'Invalid request'}, status=400)
+            data = json.loads(request.body)
+            game_id = int(data.get('game_id'))
+            if game_id < 1:
+                return JsonResponse({'Error': 'Invalid game_id'}, status=400)
             game = PongService.delete_game(game_id, user)
             return JsonResponse(game, status=200)
         except Exception as e:
             return JsonResponse({'Error': str(e)}, status=400)
         
 @jwt_auth_required()
-def get_game_by_id(request,user : User, game_id) -> JsonResponse:
+def get_game_by_id(request,user : User, game_id: int) -> JsonResponse:
     """
     Returns the game by id
     @param request: HttpRequest
@@ -73,6 +79,8 @@ def get_game_by_id(request,user : User, game_id) -> JsonResponse:
     @return: JsonResponse with game schema
     """
     try:
+        if game_id < 1:
+            return JsonResponse({'Error': 'Invalid game_id'}, status=400)
         game = PongService.get_game_by_id(game_id)
         return JsonResponse(game, status=200)
     except Exception as e:
@@ -88,12 +96,12 @@ def get_user_games(request, user: User, id: int) -> JsonResponse:
     @param id: int
     @return: JsonResponse with game schema
     """
-    type = str(request.GET.get('type', ''))
-    skip = int(request.GET.get('skip', 0))
-    limit = int(request.GET.get('limit', 10))
-    if type == '':
-        type = 'all'
     try:
+        type = str(request.GET.get('type', 'all'))
+        skip = int(request.GET.get('skip', 0))
+        limit = int(request.GET.get('limit', 10))
+        if skip < 0 or limit < 0:
+            return JsonResponse({'Error': 'Invalid skip or limit'}, status=400)
         games = PongService.get_user_games(id, type, skip, limit, user)
         return JsonResponse(games, status=200, safe=False)
     except Exception as e:
